pyproject.toml
[tool.poetry]
name = "multiworldhostservice"
version = "0.1.0"
description = ""
authors = ["Thomas Prescott <tcprescott@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.8.4"
quart = "^0.18.3"
websockets = ">=8.1,<9.0"
aenum = "^3.1.11"
fast-enum = "^1.3.0"
python-bps-continued = "^7"
colorama = "^0.4.6"
aioconsole = "^0.6.0"
pyyaml = "^6.0"
requests = "^2.28.2"
aerich = "^0.7.1"
tortoise-orm = {extras = ["asyncmy"], version = "^0.19.3"}
shortuuid = "^1.0.11"


[tool.poetry.group.dev.dependencies]
ipython = "^8.11.0"


[tool.aerich]
tortoise_orm = "migrations.tortoise_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
